feat: refactor GoogleMap component and update image handling in Navbar and Footer
